ISIS-2726: moves InteractionTracker to applib
... there's no reason not to expose this functionality, and it enables simplifying SudoService
