+++ core/runtime/src/main/java/org/apache/isis/core/runtime/events/_Xray.java
@@ -21,7 +21,7 @@ package org.apache.isis.core.runtime.events;
 import org.apache.isis.applib.services.confview.ConfigurationViewService;
 import org.apache.isis.commons.internal.debug.xray.XrayDataModel;
 import org.apache.isis.commons.internal.debug.xray.XrayUi;
-import org.apache.isis.core.interaction.session.InteractionTracker;
+import org.apache.isis.applib.services.iactnlayer.InteractionTracker;
 import org.apache.isis.core.security.util.XrayUtil;
 
 import lombok.val;
